KIO - Adding Dimens to Dashboard Categories update 3/3
The beginning to an end.. "exposed"..

Adding  Dimens to give ability for base overlays to override existing layers,
by adding corners to dashboard categories.

** For example:
User wants to apply an overlay that just modifies the dashboard_categories corners (radius), but wants to keep existing theme "color" and "style" this overlay would just 	 override the radius of the dashboard_categories without affecting applied theme.

**
android:radius does not reference a dimens in dimens.xml because if a user applies a
theme that modifies just the android:radius and they do not want that style.. The
below dimens will override that.

**
As always does not affect stock, nor does it affect old themes.